update user and order api

+++ app/Http/Controllers/Api/OrderController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
+use App\Http\Resources\OrderResource;
 use App\Models\Address;
 use App\Models\Order;
 use App\Models\Prescription;
@@ -15,18 +16,23 @@ class OrderController extends Controller
     public function index()
     {
         $user = Auth::user();
-        $allOrders = Order::where('user_id', $user->typeable->id)->get();
-        
-        // return view('order.index', ['orders' => $allOrders]);  
+        $allOrders = Order::where('end_user_id', $user->typeable->id)->get();
+        if ($allOrders) {
+            return OrderResource::collection($allOrders);
+        } else {
+            return "no addresses";
+        }
+        // return view('order.index', ['orders' => $allOrders]);
     }
 
-    public  function show($id)
+    public function show($id)
     {
         $order = Order::find($id);
         // return view('order.show', ['order' => $order]);
+        return new OrderResource($order);
     }
 
-    public  function store()
+    public function store()
     {
         $user = Auth::user();
         $order = new Order();
@@ -36,7 +42,7 @@ class OrderController extends Controller
 
         $order->creator_type = 'user';
         $user = $user->typeable;
-        $order->user()->associate($user);
+        $order->endUser()->associate($user);
         $order->save();
 
         $Prescriptions = request()->file("Prescriptions");
@@ -59,41 +65,53 @@ class OrderController extends Controller
         ]);
     }
 
-    public  function update($id)
+    public function update($id)
     {
         $user = Auth::user();
         $order = Order::find($id);
-        $order->update([
-            'is_insured' => request()->radio,
-            'address_id' => request()->address
-        ]);
-
-        $prescriptions = request()->file("Prescriptions");
-        if ($prescriptions) {
-            $old_prescriptions = $order->prescriptions()->get(); //get all prescriptions of this order
-            foreach ($old_prescriptions as $prescription) {
-                if (Storage::exists($prescription->prescription)) {
-                    Storage::delete($prescription->prescription); //delete from storage
-                    //delete record
-                }
-                $prescription->delete(); //delete all old prescriptions to add new ones
-            }
-            foreach ($prescriptions as $Prescription) { //save new ones in storage
-                $paths[] = Storage::putFileAs(
-                    'public/prescriptions',
-                    $Prescription,
-                    $Prescription->getClientOriginalName()
-                );
-            }
-            foreach ($paths as $path) {  //create new instances
-                Prescription::create([
-                    'order_id' => $id,
-                    'prescription' => $path
+        // dd($user->hasRole(['end-user']));
+        if($order->status=="new") {
+             //if end user
+                $order->update([
+                    'is_insured' => request()->radio,
+                    'address_id' => request()->address
                 ]);
-            }
+
+                $prescriptions = request()->file("Prescriptions");
+                if ($prescriptions) {
+                    $old_prescriptions = $order->prescriptions()->get(); //get all prescriptions of this order
+                    foreach ($old_prescriptions as $prescription) {
+                        if (Storage::exists($prescription->prescription)) {
+                            Storage::delete($prescription->prescription); //delete from storage
+                            //delete record
+                        }
+                        $prescription->delete(); //delete all old prescriptions to add new ones
+                    }
+                    foreach ($prescriptions as $Prescription) { //save new ones in storage
+                        $paths[] = Storage::putFileAs(
+                            'public/prescriptions',
+                            $Prescription,
+                            $Prescription->getClientOriginalName()
+                        );
+                    }
+                    foreach ($paths as $path) {  //create new instances
+                        Prescription::create([
+                            'order_id' => $id,
+                            'prescription' => $path
+                        ]);
+                    }
+                }
+                // if ($user->hasRole('end-user')) {
+                    return response()->json([
+                        'message' => 'order updated successfully'
+                    ]);
+                // }
+            
+
+        } else {
+            return response()->json([
+                'message' => 'you can not edit the order because it is '.$order->status.''
+            ]);
         }
-        return response()->json([
-            'message' => 'order updated successfully'
-        ]);
     }
 }
